Breaks down TestModule's bindings into smaller modules (more to come)
