+++ kifi-backend/modules/common/app/com/keepit/shoebox/ShoeboxServiceClient.scala
@@ -350,22 +350,3 @@ class ShoeboxServiceClientImpl @Inject() (
   }
 
 }
-
-trait ShoeboxServiceClientModule extends ScalaModule
-
-case class ShoeboxServiceClientImplModule() extends ShoeboxServiceClientModule {
-  def configure() {}
-
-  @Singleton
-  @Provides
-  def shoeboxServiceClient (
-      client: HttpClient,
-      cacheProvider: ShoeboxCacheProvider,
-      serviceDiscovery: ServiceDiscovery): ShoeboxServiceClient = {
-    new ShoeboxServiceClientImpl(
-      serviceDiscovery.serviceCluster(ServiceType.SHOEBOX),
-      current.configuration.getInt("service.shoebox.port").get,
-      client, cacheProvider)
-  }
-
-}
