+++ kifi-backend/modules/common/app/com/keepit/shoebox/ShoeboxServiceClientModule.scala
@@ -1,0 +1,27 @@
+package com.keepit.shoebox
+
+import com.google.inject.{Provides, Singleton}
+import com.keepit.common.net.HttpClient
+import com.keepit.common.zookeeper.ServiceDiscovery
+import com.keepit.common.service.ServiceType
+import play.api.Play._
+import net.codingwell.scalaguice.ScalaModule
+
+trait ShoeboxServiceClientModule extends ScalaModule
+
+case class ProdShoeboxServiceClientModule() extends ShoeboxServiceClientModule {
+  def configure() {}
+
+  @Singleton
+  @Provides
+  def shoeboxServiceClient (
+    client: HttpClient,
+    cacheProvider: ShoeboxCacheProvider,
+    serviceDiscovery: ServiceDiscovery): ShoeboxServiceClient = {
+    new ShoeboxServiceClientImpl(
+      serviceDiscovery.serviceCluster(ServiceType.SHOEBOX),
+      current.configuration.getInt("service.shoebox.port").get,
+      client, cacheProvider)
+  }
+
+}
