+++ kifi-backend/modules/common/test/com/keepit/classify/FakeDomainTagImporterModule.scala
@@ -1,0 +1,10 @@
+package com.keepit.classify
+
+import com.google.inject.{Provides, Singleton}
+
+case class FakeDomainTagImporterModule() extends DomainTagImporterModule {
+
+  @Singleton
+  @Provides
+  def domainTagImportSettings: DomainTagImportSettings = DomainTagImportSettings(localDir = "", url = "")
+}
