+++ kifi-backend/modules/common/test/com/keepit/common/healthcheck/BabysitterTest.scala
@@ -1,34 +1,18 @@
 package com.keepit.common.healthcheck
 
 import org.specs2.mutable.Specification
-import com.keepit.inject._
-import com.keepit.model.ExperimentTypes.ADMIN
-import com.keepit.social.SecureSocialUserService
-import com.keepit.test.EmptyApplication
 import com.keepit.test.FakeClock
-import play.api.libs.concurrent.Execution.Implicits._
-import play.api.libs.json.JsValue
-import play.api.test.Helpers._
-import play.api._
-import play.api.libs.concurrent.Akka
-import com.keepit.common.logging.Logging
-import play.api.libs.concurrent.Akka
-import com.keepit.inject._
 import org.joda.time.DateTime
-import akka.actor.Scheduler
-import play.api.libs.concurrent.Execution.Implicits.defaultContext
-import scala.concurrent.duration._
 import scala.concurrent.duration._
 import com.keepit.common.time._
 import com.keepit.test.TestInjector
 import com.keepit.test.BabysitterModule
-import com.keepit.test.FakeSchedulerModule
 
 class BabysitterTest extends Specification with TestInjector {
 
   "Babysitter" should {
     "do nothing if code executes quickly" in {
-      withCustomInjector(BabysitterModule(), FakeSchedulerModule())  { implicit injector =>
+      withCustomInjector(BabysitterModule())  { implicit injector =>
         inject[Babysitter].watch(BabysitterTimeout(Duration(1, "seconds"), Duration(1, "seconds"))) {
           // So fast!
         }
