+++ kifi-backend/modules/common/test/com/keepit/common/net/FakeHttpClientModule.scala
@@ -8,3 +8,5 @@ case class FakeHttpClientModule(requestToResponse: PartialFunction[String, FakeC
     bind[HttpClient].toInstance(new FakeHttpClient(Some(requestToResponse)))
   }
 }
+
+
