+++ kifi-backend/modules/common/test/com/keepit/model/TestSliderHistoryTrackerModule.scala
@@ -1,0 +1,15 @@
+package com.keepit.model
+
+import com.google.inject.{Singleton, Provides}
+import com.keepit.common.db.slick.Database
+
+case class TestSliderHistoryTrackerModule() extends SliderHistoryTrackerModule {
+
+  def configure() {}
+
+  @Provides
+  @Singleton
+  def sliderHistoryTracker(sliderHistoryRepo: SliderHistoryRepo, db: Database): SliderHistoryTracker =
+    new SliderHistoryTrackerImpl(sliderHistoryRepo, db, -1, -1, -1)
+
+}
