+++ kifi-backend/modules/search/app/com/keepit/dev/SearchDevModule.scala
@@ -1,19 +1,19 @@
 package com.keepit.dev
 
 import com.keepit.search._
-import com.keepit.module.{LocalDiscoveryModule, DevActorSystemModule}
-import com.keepit.shoebox.ShoeboxServiceClientImplModule
+import com.keepit.shoebox.ProdShoeboxServiceClientModule
 import com.keepit.common.cache.SearchCacheModule
 import com.keepit.common.SearchBrowsingHistoryModule
 import com.keepit.common.SearchClickHistoryModule
 import com.keepit.search.SearchConfigModule
 import com.keepit.common.cache.HashMapMemoryCacheModule
 import com.keepit.social.RemoteSecureSocialModule
-import com.keepit.common.service.ServiceType
 import com.keepit.common.healthcheck.HealthCheckProdModule
 import com.keepit.common.store.SearchDevStoreModule
 import com.keepit.common.net.ProdHttpClientModule
 import com.keepit.inject.FortyTwoModule
+import com.keepit.common.actor.DevActorSystemModule
+import com.keepit.common.zookeeper.DevDiscoveryModule
 
 case class SearchDevModule() extends SearchModule(
 
@@ -21,11 +21,11 @@ case class SearchDevModule() extends SearchModule(
   fortyTwoModule = FortyTwoModule(),
   cacheModule = SearchCacheModule(HashMapMemoryCacheModule()),
   secureSocialModule = RemoteSecureSocialModule(),
-  shoeboxServiceClientModule = ShoeboxServiceClientImplModule(),
+  shoeboxServiceClientModule = ProdShoeboxServiceClientModule(),
   clickHistoryModule = SearchClickHistoryModule(),
   browsingHistoryModule = SearchBrowsingHistoryModule(),
   actorSystemModule = DevActorSystemModule(),
-  discoveryModule = LocalDiscoveryModule(ServiceType.DEV_MODE),
+  discoveryModule = DevDiscoveryModule(),
   healthCheckModule = HealthCheckProdModule(),
   storeModule = SearchDevStoreModule(),
   httpClientModule = ProdHttpClientModule(),
