+++ kifi-backend/modules/search/app/com/keepit/search/SearchModule.scala
@@ -5,11 +5,12 @@ import com.keepit.social.SecureSocialModule
 import com.keepit.shoebox.ShoeboxServiceClientModule
 import net.codingwell.scalaguice.ScalaModule
 import com.keepit.model.{BrowsingHistoryModule, ClickHistoryModule}
-import com.keepit.module.{ConfigurationModule, ActorSystemModule, DiscoveryModule}
 import com.keepit.common.healthcheck.HealthCheckModule
 import com.keepit.common.store.StoreModule
 import com.keepit.common.net.HttpClientModule
-import com.keepit.inject.FortyTwoModule
+import com.keepit.inject.{ConfigurationModule, FortyTwoModule}
+import com.keepit.common.actor.ActorSystemModule
+import com.keepit.common.zookeeper.DiscoveryModule
 
 abstract class SearchModule(
