+++ kifi-backend/modules/shoebox/app/com/keepit/dev/ShoeboxDevModule.scala
@@ -3,29 +3,28 @@ package com.keepit.dev
 import com.keepit.common.mail._
 import com.keepit.shoebox._
 import com.keepit.common.db.SlickModule
-import com.keepit.module.DevActorSystemModule
 import com.keepit.shoebox.ShoeboxBrowsingHistoryModule
 import com.keepit.common.social.SocialGraphImplModule
 import com.keepit.model.ProdSliderHistoryTrackerModule
 import com.keepit.learning.topicmodel.DevTopicModelModule
 import com.keepit.realtime.ShoeboxWebSocketModule
 import com.keepit.scraper.ScraperImplModule
-import com.keepit.search.SearchServiceClientImplModule
+import com.keepit.search.ProdSearchServiceClientModule
 import com.keepit.shoebox.ShoeboxClickHistoryModule
 import com.keepit.shoebox.ShoeboxDbInfo
 import com.keepit.common.cache.ShoeboxCacheModule
 import com.keepit.classify.DevDomainTagImporterModule
 import com.keepit.common.cache.HashMapMemoryCacheModule
-import com.keepit.module.LocalDiscoveryModule
 import com.keepit.shoebox.UserIndexModule
 import com.keepit.social.ShoeboxSecureSocialModule
 import com.keepit.common.analytics.DevAnalyticsModule
 import com.keepit.common.crypto.ShoeboxCryptoModule
-import com.keepit.common.store.{ShoeboxDevStoreModule, ShoeboxProdStoreModule}
-import com.keepit.common.service.ServiceType
+import com.keepit.common.store.ShoeboxDevStoreModule
 import com.keepit.common.healthcheck.HealthCheckProdModule
 import com.keepit.common.net.ProdHttpClientModule
 import com.keepit.inject.FortyTwoModule
+import com.keepit.common.actor.DevActorSystemModule
+import com.keepit.common.zookeeper.DevDiscoveryModule
 
 case class ShoeboxDevModule() extends ShoeboxModule(
 
@@ -34,14 +33,14 @@ case class ShoeboxDevModule() extends ShoeboxModule(
   fortyTwoModule = FortyTwoModule(),
   cacheModule = ShoeboxCacheModule(HashMapMemoryCacheModule()),
   secureSocialModule = ShoeboxSecureSocialModule(),
-  searchServiceClientModule = SearchServiceClientImplModule(),
+  searchServiceClientModule = ProdSearchServiceClientModule(),
   clickHistoryModule = ShoeboxClickHistoryModule(),
   browsingHistoryModule = ShoeboxBrowsingHistoryModule(),
   mailModule = DevMailModule(),
   cryptoModule = ShoeboxCryptoModule(),
   storeModule = ShoeboxDevStoreModule(),
   actorSystemModule = DevActorSystemModule(),
-  discoveryModule = LocalDiscoveryModule(ServiceType.DEV_MODE),
+  discoveryModule = DevDiscoveryModule(),
   healthCheckModule = HealthCheckProdModule(),
   httpClientModule = ProdHttpClientModule(),
