+++ kifi-backend/modules/shoebox/app/com/keepit/shoebox/ShoeboxProdModule.scala
@@ -2,7 +2,7 @@ package com.keepit.shoebox
 
 import com.keepit.common.cache.{EhCacheCacheModule, MemcachedCacheModule, ShoeboxCacheModule}
 import com.keepit.social.ShoeboxSecureSocialModule
-import com.keepit.search.SearchServiceClientImplModule
+import com.keepit.search.ProdSearchServiceClientModule
 import com.keepit.common.db.{DbInfo, SlickModule}
 import com.keepit.scraper.ScraperImplModule
 import com.keepit.common.social.SocialGraphImplModule
@@ -17,17 +17,18 @@ import com.keepit.classify.ProdDomainTagImporterModule
 import scala.slick.session.{Database => SlickDatabase}
 import play.api.db.DB
 import play.api.Play
-import com.keepit.module.{ProdDiscoveryModule, ProdActorSystemModule}
 import com.keepit.common.healthcheck.HealthCheckProdModule
 import com.keepit.common.net.ProdHttpClientModule
 import com.keepit.inject.FortyTwoModule
+import com.keepit.common.actor.ProdActorSystemModule
+import com.keepit.common.zookeeper.ProdDiscoveryModule
 
 case class ShoeboxProdModule() extends ShoeboxModule(
   // Common Functional Modules
   fortyTwoModule = FortyTwoModule(),
   cacheModule = ShoeboxCacheModule(MemcachedCacheModule(), EhCacheCacheModule()),
   secureSocialModule = ShoeboxSecureSocialModule(),
-  searchServiceClientModule = SearchServiceClientImplModule(),
+  searchServiceClientModule = ProdSearchServiceClientModule(),
   clickHistoryModule = ShoeboxClickHistoryModule(),
   browsingHistoryModule = ShoeboxBrowsingHistoryModule(),
   mailModule = ProdMailModule(),
